Started rationalizing Trading API and impl class locations
